Solucionado problema con los colores. El slider coge el mayor valor de las facturas.

+++ app/src/main/java/com/example/a/model/Factura.kt
@@ -10,7 +10,7 @@ data class Factura(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
     @SerializedName("descEstado") val pendientePago: String,
-    @SerializedName("importeOrdenacion") val cantidad: Double,
+    @SerializedName("importeOrdenacion") val cantidad: Float,
     @SerializedName("fecha") val fecha: String
 
 )
